refactor: rename reducer
