+++ src/components/ImageMapEditor/useImageMapEditorReducer.ts
